requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
pyTelegramBotAPI==4.7.1
requests==2.28.1
telebot==0.0.4
urllib3==1.26.12
psycopg2-binary==2.9.9
